requirements.txt
numpy>=1.13.0
Biopython
click